Refactor OpenAI fallback handling into dedicated module
